attached messageQ to a generic Internet class this allows message passing in Multicast streams in case of timeout errors.
Other changes include fixing a silly bug in reading the stats from the
MPEG module, not allowing zero as a message handle.

Most importantly the list class now has a semaphor.

SetServer and OpenStream API's reflect the passing of a generic internet
address in set server.  As such Open stream opens one or more assets on a
given server.  If that server is a multicast server then it opens its implicit
stream.