included the error code returned from the server in server error.
Removed Errors before closing a stream this allows the operation of
the close stream.

Semaphored the server class.  This is due to allowing internet events
whenever, that means without semaphoring an asset request response could
be interupted by an internet event and generally getting confused.  this
action is now barred.

created a multicast server based on host ordered address, this was ignored
and the address supposed not to be multicast, therefore an attempt was made
to create a VOD server out of it.

replaced the -1 preplay ack as it was causing havoc.