bit 4 in R0 in set server could under some conditions be deleted before testing to see if it is set, fixed.
-from in the lanTV command didn't work as I tested r4 to be 0 not R0 as well
this meant that I would try to find the server with the from time as its id.
fixed.

it was possible to delete the server object calling setserver (delete) while
that server was being used in a stream. fixed.

It was possible to move into a play state from a ff state if going through
pause or step.  This has been fixed.

I have also put code in to ensure that the memory used for error messages is
correctly handled.