in the server control class I removed the wait for play request response this was commented out anyway.
in the sink class moved the call to the reset stream to the header in order
for it to be placed inline.

in the stream class, where the majority of changes has occurred, I have
changed the highspeed file into a enumerated type retaining knowledge of
the state.  this is then used to determine if there is a need to change the
continuity of the stream.

If the continuity of the stream is changed then the stream is reset.  This
resetting was traditionally in the application, but I feel it is done with
a greater degree of control from the VSP module.

I also do not send errors when the state is stop this is due to the use
of the cannot delete shared memory error usage.