Fixed a bug in List where old errors were not ignored as should be this stopped the use of opening a server once a server couldn't be found.
implemented the standby service call, this adds the *vspdisconnect command
that will disconnect all existing connections.  This calls the same code that
the service call calls.

fixed a potential problem in the returning of error blocks.

Added a debug word to isolate areas of debug such as entry exit points.  This
accounts for the majority of changes to files.  It adds the inline
if (Debug(flag)) {X}
This will do X if that flag is set in the debug word.