I have made changes to the callback handling to make them more timely and exact.  This follows a bug report for the STB going into a state of not requesting more information.  This was later found to be due to the MPEG file and the amount of buffers handed to it. This is not fully resolved.
A few error handling errors have been fixed most importantly the deleting of
the default server if it could not be connected to, and raising an error.

there is one limitation that is not resolved that is bug in this release it
is the stomping on the messageQ error pointer this will put the STB into a
corrupt state.

Start/end notifications are implemented as per the PC client in this it ignores
the block number but allows timeouts after receiving information.