Implimented blanking between scenes.  This includes a system variable enabling the switching of this option off.
Fixed a scenelist problem where errors weren't being copied to the interface.

Fixed a problem where previous streams would set the message sent flag not
enabling the start of stream to be sent.

Most importantly instead of timing out on a block I assume it is deleted and
move on.  If three blocks are deleted in such a manor then I assume an error
in connection.  This should be done from the server side.