Summary: 	Update of TarExtend to improve support for archiving builds. Now at Version 0.03
Details:
	Altered the dearchival script creation routine to get the tar and filecrc
tools from <dearchiver_tool$dir> instead of from the library path. This means
that the system variable can be made to point to a set store (say a 'Tools'
directory on a CD) and use those instead of whatever old versions may be in
the library path.

	Added -crc and -single to the filecrc scripts to make them work
better with older versions of FileCRC.

Version 0.21. Tagged as 'Library-0_21'