changes since 1.23 are ;
* Reset stream done when a stream changes play file continuity has been moved
from the test token code to the code that clears the bufferswait flag.

* step play now pauses the mpeg player instead of 'drying up' this resolves issues
surrounding break-up in step play.

* the timeout for the repeating of a blocks naks is half the inettimeout used for the
timeout of blocks.

* step play changes the mode to normal play when stepping through the stream.

* the bufferswait flag is cleared on the receipt of a start/end notification this allows
the playing of small portions of the stream near the end.

* the block timeout now takes the number of blocks between the current time and the time
when this block is expected to turn up in to account when assigning a timeout to a new
block.