Summary: 	Fixed a problem where repeatedly pressing the 'connect' and 'logout' buttons could cause NCMAlite to go onto an inconsistent state on ethernet builds.
Detail:
	If you press the 'logout' button repeatedly when browsing could cause
NCMAlite's state machine to get into such a state where it is in the
'connect' state yet still displays the web page the user was browsing when
connect was pressed.
	Kluged a fix to this by handling Trans_Logout in State_Connect and
making it open up the 'connect' screen. Hence if it ever does get into this
incosistent state it should automatically open up the correct screen.
	Also did a little optimistaion in the call to NetTime_Kick thanks to
Kevin noticing that the eisting code is highly non-opitmal.

Admin:
 	Tested as a softload on an MPT NC (ethernet and modem interfaces).

Version 0.45. Tagged as 'NCMAlite-0_45'