Summary: 	Fixed fault report 988 (startup sound is interrupted occasionally).
Detail:
 	Occasionally the startup sound would be interrupted when booting the
box. To fix this, have increased the size of the FilePlay buffers in the UK
and Japanese !boot.!run files from 12K to 20K. This appears to have fixed the
problem; extensive rebooting with the increased buffers did not show up the
fault.

Admin:
	Tested in a development build on an MPT NC (modem and ethernet
interfaces)

Version 0.54. Tagged as 'MPTUI-0_54'