Summary:  	Fixed a bug in srcbuild where a fatal error caused the SrcBuild$ReturnCode system variable to be set to 1 (representing a successful build) instead of 0 (representing a failed build).
Details:
	Whether the build was successful or not is determined by whether an
error is present in the error table. If the error was flagged as being a
fatal one, then exit() was being called *before* the error was added to the
error table. Hence when the atexit()-registered routine was called, there were
no errors in the error table and the system variable was being set as 1.
	Fixed by moving the exit() call that occurs in the error() routine
past the code that enters the error into the error table.

Admin:
	Tested in a deliberately-botched MPT build.

Version 0.14. Tagged as 'srcbuild-0_14'