Summary: 	Added second and third stages of new functionality requested by MPT.
Detail:
	If a default PIN of 9999 was entered, it now opens a new screen,
connect-np, which does not have a 'change PIN' button, instead of the normal
connect screen. This prevents the user from being able to change the PIN on
smartcards with a second efault PIN enabled. This was done by creating a new
routine, select_connect_screen_to_display(), which openes the relevant
screen. A call to this routine replaces all the existing
ncma_url_open(URL_PATH(connect.html)) calls in the NCMA state machine.
	Also, if the new behaviour is enabled, you can no longer change the
PIN on a smartcard to be the second default PIN.

Admin:
	Tested as a softload on an MPT Ethernet NC. Had four cups of coffee.
Work in progress.

Version 0.49. Tagged as 'NCMAlite-0_49'