Summary: 	Created a new ncma command, open_connect_screen, which can be used to tell ncma to open the correct connect screen.
Detail:
	As part of the ongoing work for MPT, it was discovered that NCOptions
and ISPcapture both try and open the ncma connect screen directly. This broke
the case in MPT where we wanted to use an alternate connect screen to 'hide'
the 'change PIN' button. This command can now be called by NCOptions and
ISPCapture to open the connect screen.

Admin:
	Tested as a softload on an MPT Ethernet NC. Drunk three more cups of
coffee and had two packs of crisps.

Version 0.50. Tagged as 'NCMAlite-0_50'