The GNU General Public License.
This license applies to all software sources found within this directory
tree.

To summarise the key points: it does not place any restrictions on the
use to which we may put these sources and it does not oblige us to
release these sources or otherwise make them available externally
unless we choose to distribute the resulting binaries externally.