Summary: 	Altered code so that pressing standby when on options screen, smartcard screen or dialling no longer takes the user back to the 'connect' screen if the DISALLOW_USER_CONFIRMATION_OF_STANDBY preprocessor flag is set.
Detail:
	Kevin's previous change ignored the 'Service_Shutdown' call if the
flag above was set. This prevented the standby button from taking the user
back to the 'connect' screen when broswing, as required by the Lazarus
project. However, when on other screens (eg. the options screens) it wouls
still go back to the 'connect' screen when emerging from Standby.
	Fixed by simply masking out the 'Service_ShutdownComplete' flag if
the flag was set.
	Also tidied up the code a tiny bit.

Admin:
 	Tested as a softload on an NC1 running a Lazarus build.

Version 0.55. Tagged as 'NCMAlite-0_55'