Summary: 	Added an extra transition to allow the user to close the 'invalid card' dbox.
Detail:
 	On Lazarus builds there are no smartcards. Since Lazarus builds are
built with the ISP_CAPTURE preprocessor flag undefined (to prevent the
ISPCapture screens from opening if the NVRAM bootblock is not set) then
'invalid card' dbox can be opened.
 	The 'invalid card' dbox is closed by the user removing the smartcard.
Sicne we don't have a smartcard on Lazarus builds, I had added a transition,
Trans_BadCardNak, that can be invoked from a button on the invalid card dbox
to close the dbox and go down to the connect screen and state.

Admin:
	Built.

Version 0.56. Tagged as 'NCMAlite-0_56'