Summary: 	Fixes for faults 1081, 1082 and 1083.
Detail:
 	Altered the 'invalid' screen so the pressing the button gives a
ncma:badcardnak instead of a ncma:badcardnack which was not being recognised.
(Part of fix for fault 1081).
	Altered the mappings file so that the BHS_printing_b_printer_bj-f100
now points to BJC1000 and BHS_printing_b_printer_bj-f200 now points to
BJC2000, which are the English names for these printers.
 	Altered the 'Printing' options screen so that the 'complete' button
now goes to the 'printing2' screen instead of back to the 'options' screen.

Admin:
 	Tested in a temporary build.

Version 0.16. Tagged as 'LazarusUI-0_16'