New overview document in Docs, requested by some Pace people.
Printing engine will attempt not to seperate headings from non-heading
items. If the current line is a heading and the next is not, and falls
off the bottom edge of the paper, then both lines will be carried over
to the next page. Similarly for DT and DD items. Maximum line height to
push rather than split is now altered to 2/3 of page height (was 1/2).

JavaScript fix; a URL's "protocol" value should include a ':' at the end.