Summary: ========  	Fixed bug where valid but empty NVRAM tags caused all keypresses to be interpreted as a valid key.  	Added a little more debugging information.
Detail:
=======
 	DOH!
 	If a key was held in an NVRAM tag and the tag existed but was empty
(eg. following a Power-on Reset) then an empty string was put as the string
to be matched and hence every keypress on the correct URL would be
interpreted a correct match. This has now been fixed - empty strings are
*not* valid keys. Strangely enough the old behaviour matched the spec, as the
spec did not state that empty strings were not valid keys.

Admin:
======
 	Tested as a softload on an NC1 running LazarusOmega build.

Version 0.07. Tagged as 'PassFilter-0_07'