Summary: ========  	Started adding support for the hidden phone number.
Detail:
=======
 	There is a requirement in the Lazarus project to allow dialling to
start with an alternate test phone number on the production line. This is
being supported in the Lazarus builds by having a secret PassFilter key on
the connect screen that will cause NCMA to dial a number held in an NVRAM
tag.
 	Hence have added two new NVRAM tags:
 	* NCMAHiddenPhoneNumber (15 bytes)
 	* NCMAStartHiddenDialPassword (11 bytes)
 	And have altered the !Boot.Config file to send a
ncma:ConnectWithHiddenPhoneNumber command when the key held in the
NCMAStartHiddenDialPassword tag is entered.

Admin:
======
	Work in progress.

Version 0.27. Tagged as 'LazarusUI-0_27'