Summary: 	Altered to work with HdrSrc-0_45. 	Placed under SrcCommit control.
Detail:
	HdrSrc 0.45 has had some of the macros from CVars.hdr file moved into
Global.Macros, as the CVars file was present in many places in the source
tree. Altered SoundFIle so that it no longer attempts to include it's local
copy of CVars, and altered so that it builds with the generic copy from
HdrSrc.
	Altered CMHG headers so that it now build version number and date
string from the VersionNum file.

Admin:
	Tested locally. Requires HdrSrc-0_45 to build.

Version 0.26. Tagged as 'SoundFile-0_26'