Summary:  	Rewrote the memory handling subsystems to fix a long-standing bugette.  	Added a preprocessor option, LOGOUTS_ONLY_HANGUP, that makes logout requests received whilst in Booted state only hangup.   	Added an options.h file that allows build options to be set and stops me from having to check in alterations to component files all the time.
Detail:
 	Finished off the work started in the last checkin on the memory
subsystems. This should be much more stable. Debugging builds now report all
requests and frees of memory.
	Lazarus UI builds need the 'hangup' button to do exactly that, only
hangup and *not* logout. Added a preprocessor flag to allow this behaviour.
	NCMAlite has many build options (many of which are historical in
nature), all of which are passed in on the command line at build time.
Recently the command-line length limitation of RISC OS has been reached, and
hence I have started a system using an options.h file that allows people to
define differing flags according to the UserIF.
	This is only a limited implementation of this behaviour and it will
be extended as time goes on. It could be extended to set varioables on
Machine as well as UserIF.

Admin:
 	Tested as a softload on a Lazarus IAC build.


Version 0.61. Tagged as 'NCMAlite-0_61'