Admin: 	Implemented a fix for the  bug where invalid URL's on the smartcard (eg. www.acorn.com) caused the machine to continue displaying the 'Please wait' screen.
Detail:
 	If the browser does not recognise a URL as not having a protocol that
it deals with (eg. http:// or ftp://) then it ignores the message to open the
URL. If the home URL on the smartcard is of that format then the attempt to
open the home URL fails silently, leaving the 'Please Wait' screen up.
 	The simple fix to this it to detect the bounce of the OpenURL message
and open the BadURL screen. However, since we were using
TaskModule_SendMessage to send the OpenURL message we cannot get the my_ref
field out of it to detect whether the boucned message was sent from us.
	To overcome this, it now sends the message from the task section of
NCMAlite. However, due to callback problems I trigger the actual sending of
the OpenURL message on a Wimp Poll.

Admin:
	Tested as a softload on a Lazarus IAC. There is a possibility that
other OpenURL messages sent from NCMAlite may bounce, in which case they will
bring up the BadURL screen. This may not be desired behaviour.

Version 0.66. Tagged as 'NCMAlite-0_66'