Summary: 	Added comments to ncma_cmds.c 	Fixed bug added in last checkin.
Detail:
	A bug was introduced during the ncma_url_open() alterations where I
was doing a strncpy() of a string and not including the terminator. Hence if
a larger string was overwritten by a smaller one then the string would appear
with the tail of the larger one added. This only shows up on email client
builds.

Admin:
	Tested on an email client enable build.

Version 0.68. Tagged as 'NCMAlite-0_68'