Debug target added for DDT, which includes basic TML output stuff. Anyone wanting full debuglib can add it if they wish, but should be aware of the assembler side of things when doing so.
Major change is the fix for claiming RMA space at startup even if using
dynamic areas overall. With large images, this could exceed the RMA limit
of 11MB and prevent the application starting - even with plenty of free
RAM. The fix is not perfect and is documented in detail in comments; see
ddt.s and look for label 'go'.

Checked that RAM and RAM debug builds build and that the clean rules
clean. Uses library binaof now, rather than a local copy. Dirs rule works
properly.