Summary:  	A fix for bug 1197.  	General tidying-up performed.
Detail:
	Synopsis of bug 1197: If you attempt a redial once you are already
browsing and cause the redial to fail, eg. by setting call waiting, and then
go into the options screen, any other redial will cause the options screen to
remain up. The user then becomes trapped in the options screen and the only
way out is to reboot the box. This is a long-standing bug, and is exhibited
on Funai2 boxes.
	This was caused because on a redial NCMAlite does not call
ncma_user_url() to actually open a URL page but relies on the existing page
being displayed (this was Funai 2 and MPT behaviour). Unfortunately, if you
open the options screen on a redial that becomes the current screen and the
browser never gets to open a new page even though it has connected.
	This could be fixed by calling ncma_user_url() at the end of a
redial; however, this routine opens the home URL of the user off the
smartcard which is *not* where the user was before the attempted redial
started.
	Fixed this by storing the current URL being displayed by the browser
before a redial is called. This URL is then re-opened at the end of a redial
process. This method is not perfect; if the user clicked on a link to invoke
the redial then he is taken back to the screen with the link after the redial
rather than the link that he clicked on. This is probably a minor problem
though. it could be fixed by the browser storing the link clicked on ina
system variable in a similar manner to the way that the current URL is
stored, and then I could open the URL held in that variable instead.
	Also used the opportunity to tidy up the machine.c file a bit. Added
some comments, changed the logical groupings of a few routines and data
declarations accordingly.
	Added some new routnes, misc_getenv(), misc_getenv_malloc(),
store_current_browser_url() and open_current_browser_url() as support
routines for this alteration.

Admin:
	Tested as a debug softload. Could do with some serious testing.

Version 0.71. Tagged as 'NCMAlite-0_71'