Summary: 	Fixed a bug where with recent diallers failures during the verify stage were not being detected and the box never progressed from the Verifying stage. 	Added DebugLib and StripDepnd support.
Detail:
	There is a fault detected on the Lazarus project where pulling the
line out during the Verifying stage was causing the box not to progress past
this stage. There is a similar fault (1195) that is invoked by attempting to
connect a Lazarus box to a fax machine, and this also appears to be fixed by
this version.
	This was fixed by putting a check into not_connected_svc() that
detects if the line is down whilst the module is in state_ppp_going up and
throwing a NCDialCtrl_NoServer service call if it is. The reasoning behind
this is that there is no way that the line can be down if it is attempting to
bring PPP up and that this state of affairs is therefore an error.
	DebugLib support added to allow me to check exactly what was going on.

Admin:
	Tested as a softload and in a ROM build BLaughton has verfied the fix
as being correct.

Version 1.24. Tagged as 'NCDialCtrl-1_24'