Summary: 	Implemented a better workaround for the Lexmark printer / IZipfs clash bug.
Detail:
	There is a known bug where booting with a zip drive enabled and then
choosing a Lexmark printer causes the box to stiff (see previous checkin).
This bug was noticed on Lazarus more than previous products because the
'none' printer option was set to 999, the same as the ZipFS setting.
	Fixed the bug by setting the 'none' printer configuration option in
NCOptions to 000. This means that ZipFS cannot be selected from the NCOptions
Printing screens as the only available options in the menu are 000, 016, 017,
018 & 019.
	Because having ZipFS enabled at boot time is a vital debugging aid, I
have re-enabled this behaviour by adding a PassFilter password that then
calls a BASIC file in !Boot.Utils that sets the PrinterCode NVRAM tag to 999,
allowing the zip drivers to be run at the next startup.
	Therefore I have regressed Simon's previous checking by putting the
izipfs:Auto line back into the !Run file. I have also removed all the lines
in the !Run file that allowed acces to the command line (via ctrl-shift-F12
and gos:) as requested in the Lazarus team meeting.

Admin:
	Tested in a build.

Version 0.50. Tagged as 'LazarusUI-0_50'