Added prodcommit.
This application is not complete.  Do not use except for manual
testing.  Checked in as a checkpoint only.

This program takes a "normal" diff output and generates a log
entry suitable for a commit log message for product module's file.

Eventually it will be used in the same way that srccommit is for
source code (hence the name prodcommit) where it primes the editor
buffer with the differences that have been made to the modules file
allowing the buildmaster to check over them (perhaps spot errors!)
and not have to worry about typing out complicated log messages.

The output message will list the component name, its old symbolic tag
and its new symbolic tag (where appropriate).  For an updated component,
it will check to see if it is using a srccommit generated tag and test
the revision numbers and append "(REGRESSED)" in the log file to
indicate that the component version number has been reduced.