Summary: 	Fixed fault 1258.
Detail:
	The UserInactivity module was failing to drop the line when the
webpage being displayed had a refresh period less than the timeout period and
the mouse pointer was not present.
	This was being caused by the fact the the browser changes the pointer
position on refreshes if icon highlighting is needed. These position changes
were being detected by this module and being taken as being user activity.
	Added a preprocessor option, IGNORE_POINTER_MOVEMENTS that makes the
module ignore mouse pointer movements. This masks out the error case, and is
deemed to be an acceptahle workaround for the Lazarus project.
	A better fix would be to detect OSWord 21,1 and 21,3 (i.e. changing
the pointer position) and mask out the next immediate pointer position chnage
after that.

Admin:
	Tested as a softload on a Lazarus board.

Version 0.03. Tagged as 'UserInactv-0_03'