Module used to collect data into 8K chunks and then send these to the MPEG software. This meant it would hang on to any trailing fragment of data, in some cases never sending it - "end of stream" markers in MPEG streams were thus often lost, for example.
Data is now sent as it arrives, without waiting to fill up an 8K buffer.
This appears to have no negative side-effects, though potentially, were
the MPEG software to decide to hang on to data for a while, it could lead
to UDPv running out of the 64 buffers it holds since it's no longer
filling each one to 8K capacity before moving onto the next. This has not
been observed in practice after a fair amount of testing.

Version 1.12. Tagged as 'UDPv-1_12'