Summary: 	Improvements to allow the script to work on Windows NT.
Detail:
	The script now checks for $USERNAME and sets the relevant $USER
system variable accordingly. This is the first step of making the scripts
work under NT. It also calls a generic delete_tree() routine instead of rm
-rf directly so behaviour can be different on differing platforms.
	The script now also uses the $^O Perl variable to detect the platform
that the Perl binary was compiled on. Note that we are using this to detect
what platform we are running on, which is naughty, but since at the moment we
only want to differentiate between Unices and Win NT, it will work.

Admin:
	Tested under Unix and Windows NT.