New improved version of srccommit.  This version contains some help for remote users which is significantly better than the previous gross and ugly hack for exporting the template log message from the server
Having now read the CVS documentation(!), I have modified the code which
obtains the default log message to attempt to read CVS/Template first - it
seems that when it is used client-server, the template is delivered to the
client during the checkout.  So the fiddling about trying to check it out
from the repository is completely unnecessary in the client-server case.

Documentation is a wonderful thing.  Look out for other vital changes
in scripts coming soon to make them do things "properly" instead of by
guessword and reverse engineering.

This version has been tested briefly on both a client-server operation and
a direct access operation.