Summary: 	Various little pieces of improved functionality and fixes.
Detail:
 	A feature of NCMAlite was that it relied on Watchdog to send a
smartcard cache enquiry early on in the boot sequence (i.e. on module
initialisation), which NCMAlite then took to mean a Trans_URLStart transition
change request. This was extremely yucky, and the functionality was broken in
Watchdog so that any keypress caused that attempted transition change. This
has been fixed so that the required transition change is sent when the task
part of NCMAlite initialises.
	Various new comment and debug line changes. Certain areas are now
more internally consistent.
	When a transition callback occurs, it now changes the transition via
a new routine, change_state(), that is simply there to allow debugging
routines to display state machine statistics before and after each change.

Admin:
	Tested as a softload on Omega and Lazarus hardware with relevent
build flags set.

Version 0.77. Tagged as 'NCMAlite-0_77'