Summary: 	Major reorganisation of the NCMAlite state machine.
Detail:
	The NCMAlite state machine has a series of global transitions
contained within it. These were a nuisance, particularly as a couple of them
were not global, but had exclusion cases. Have now altered the state machine
so that there are no global transitions, and all transitions are dealt with
in the relevant places in the individual states.
	Have also changed some of the state machine flags from int to bool,
to underline the fact that they are really booleans.
 	No new functionality has been added in this version; the behaviour
should be exactly as it was in the previous version.

Admin:
	Tested as a softload on Omega and Lazarus hardware with relevent
build flags set.

Version 0.78. Tagged as 'NCMAlite-0_78'