Summary: 	More tidying up and initial work towards reimplementation of a secure bootblock system.
Detail:
	More comments clarified.
	Flag post_initialisation_run changed from int to bool type.
 	SC_SWAPPING-specific variables are now in their own structure for
neatness.
 	if AUTHENTICATE_SIGNATURES is defined, then calls to
NCBootblock_Enquiry SWI always ask for the bootblock to be authenticated.
Initial support for possible future products that may require bootblock
authentication. NCMAlite needs to be altered to allow the error states from
this call do something if it expects a secure bootblock and does not get one.
Perhaps this could be either displaying the 'bad card' screen ot doing
something else, for instance trying to start an automatic download of ISP
details.
	Also note that currently NCBBnvram will *always* return that the
bootblock is authentic, as it ignores the flags that are passed in to the
NCBootblock_Enquiry SWI.

Admin:
	Tested as a softload on Lazarus hardware with relevent build flags
set.

Version 0.79. Tagged as 'NCMAlite-0_79'