Summary: 	Each component can now havw a -nosources flag.
Detail:
 	There is a requirement for a customer to get a binary-only build
where they can change resources and do their own builds. To allow this we
will be providing them with a prebuilt disc image for the project with almost
all the sources deleted.
	To allow them to build images, have added the functionality to
SrcBuild whereby adding a -nosources flag to a component in the components
file causes clean, export_hdrs, export_libs and rom phases to have no effect.
This allows the client to build with resources, install_rom and join phases
and thus get an image with any changed resources in.
 	This change should only affect components that have a -nsources flag
in the component file.

Admin:
	Tested in a Lazarus build.

Version 0.18. Tagged as 'srcbuild-0_18'