Summary: 	* Initial support for multiple territories in one build.
Admin:
	For the Bethany project we need to be able to build multiple resource
(territory) sets into one ROM build. The system variable LocaleList is read
and, if set, is parsed for a comma-seperated list of countries. The resources
phase is then called once for each of these countries. If the LocaleList flag
is unset then is should build the territory as set in the Locale in the Env
file.
	If the LocaleList flag is set then after the resources phase has been
run there should be one directory in the Internet.Messages path for each
territory mentioned in the LocaleList.
	To allow this to work the Common and Env files also need altering. If
the Env file does not have a LocaleList then it should have en explicit Unset
LocaleList. The common file also needs to have every *Set that references
<Locale> to be changed to a *SetMacro.

Admin:
	Work in progress for Bethany project. Only partially tested. A binary
checkin will follow after more work and testing.

Version 0.19. Tagged as 'srcbuild-0_19'