Global source change to 64-wide comments and Pace (C) messages as directed by the coding style recommendations. All header files protected against multiple inclusion. All external headers included via <> rather than "". All library based inclusions, where appropriate, included via. LibraryName/HeaderName.h rather than a tortuous -I path.
Further work, largely on the Makefile, includes continuing changes
to fit better in current build environments (particularly with
respect to internationalisation), fixing the ROM build, and so-on.

Don't build this version. The beginnings of a move to DebugLib are
in place but incomplete and the fetcher has a delay loop in it as
part of debugging a strange load balancing problem. Only the
!PhoenixJ build has up to date resources.