Summary: 	Fixed an 'OflaOfla bug.' 	Resource directory ordering changed.
Detail:
	Fixed a bug where directories called 'oflaofla...' were being created
in the Install directory. Srcbuild iterates through all the components in the
components file, and attempts to create the relevant install directory.
Unfortunatley, many components (e.g. HdrSrc) do not have an install directory
in the ModuleDB. This was causing a string to be printed with a NULL pointer,
causing oflaofla to be created as part of the directory leaf.
	As part of the work to allow multiple territories to be present in
one build, we are altering the resource directories so they are now held in the
following manner:
	Used to be : Messages.Resources.<locale>.<build>
	Is now :     Messages.Resources.<build>.<numeric terriory list>.<locale>
	Where the numeric territory list is a list of all the territories
held in the resources directory, eg. 000106 for Common, UK and France. The
Common directory holds all resources that are common to the other territories
and is generated by the Messages module. Note that if there are three
territories, then there will be four directories held in resources, one for
each territory and one for the Common directory.

Admin:
	Used in some Lazarus builds
	Very much work in progress. Use at your peril. Being checked in early
due to the anticiapted CVS server changes.
	Obbiously when you use this latest version, you will need to run the
resources phase because the formatting has changed.

Version 0.20. Tagged as 'srcbuild-0_20'