Summary: 	Yet more magic ways of fitting even more territory resources into one build.
Detail:
	We now need to fit five resource sets into the first Bethany ROM
image. The previous simple scheme that I implemented did not allow this, so
the ResCommon utility has been extensively overhauled to detect more
commonality between resource sets and hence reduce the ROM requirements for
the resources.
	For instance if there are three files, each containing a byte as
follows:
	File:	1	2	3
	Data:	A	B	A
	Then files 1 and 3 will be common, whilst 2 would have it's own,
individual, version of the file.
	This process is undertaken for all files in the resource sets. For a
Bethany build with five territories (and hence resource sets) this parsing
takes about seven minutes to calculate all commonality and copy the files
into the correct processed directories on a RiscPC with RapIde interface.

Admin:
	Tested in Bethany builds with 3 and 5 different resource sets, and in
Lazarus builds with 1 resource set. This all appears to work correctly, with
please use with caution; the first time you do a build with this version,
check the resources in the build to ensure that they at least look correct
with nothing missing.
	There are some limitations of the current system; for more detais of
these please see the 'ToDo' comments at the end of the ResCommon sources.

Version 0.96. Tagged as 'Messages-0_96'