Summary: 	ResCommon now allows resource sets *not* to have UK in them.
Detail:
	On a delete-poweron, the default territory is set to UK (01). On
startup the messages module looks at all the registered ResourceFS blocks and
uses the ones containing the currently territory.
	Unfortunately if the territories in the build do not include the UK
resource set, this means that no resources are allocated, and hence the
machine does not boot up in varies weird and wonderful ways (usually by the
WIMP failing to initialise).
	To overcome this limitation, if the territory list in the ROM does
not include UK, then the UK territory code is added to the first territory's
resource blocks. This means that the country code 01 is present in the ROM,
but it does not point to the UK territory. This allows the ROM to boot up.
	The code in ResCommon and ScanRes have been tidied.
	Added a file explaining how the resource differencing within
ResCommon works into a Docs directory.

Admin:
	Tested in Bethany and Lazarus builds. Builds containing UK resource
blocks still work, and ones without it work, and default to the first
territory in the territory list after a delete-poweron.

Version 0.98. Tagged as 'Messages-0_98'