Summary: Update of Funai 7 and Funai 8 resources.
Detail:
- Resources.NC2.Japan.Fun7JP
This directory contains updated Sprites, Res and Alt.Res files. The
effect of these is to restore the toolbars to the layout used in Funai
2/3. The sprites for the logout button have been changed to the new look
(without the 'Off' text).

- Resources.NC2.Japan.Fun8JP
This directory contains updated Sprites, Res and Alt.Res files. The Res
files have been changed to restore the correct alignment of the progress
indicator. Note that the reason this went wrong is that these Res files
contain gadgets aligned on pixel boundaries. If you edit these files you
*must* set ResEd$Options to "-align 2" otherwise ResEd will unhelpfully
go along realigning everything to 2 pixel boundaries (is there a good
reason why this is the default behaviour?) In the Sprites file the
sprites 'animt' and 'animts' have been altered to make remove the
problem of the progress indicator appearing with some of the 'sky' above
the rotating globe missing.

- Resources.Translate
Various files have been updated to allow the background color of input
boxes to be different from the window background colour. To do this a
new tag 'NCFM_inputbgcol' has been added to all the 'Mappings' files,
and this is now used in all the pages that have input boxes ('jsprompt',
'password', 'sslpin' and 'url'). Note that the only mapping file where
this is different from the window background colour is
'Translate.NC2.Japan.Fun8JP.Mappings'. This file isn't included because
it hasn't changed - I had put this new tag in the mapping file I sent
you last time and also in the 'Translate.NC2.Fun8JP.html.url' file, but
had forgotten to put it in the other mapping files and update the common
tagged html files.

Admin:
Untested.

Version 3.49. Tagged as 'NCFrescoRO-3_49'